requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
stripe==7.4.0
Pillow==10.1.0
django-cors-headers==4.3.1
django-filter==23.5
python-decouple==3.8
django-timezone-field==6.1.0
PyJWT==2.8.0
django-redis==5.4.0
reportlab==4.0.7
gunicorn==21.2.0
dj-database-url==3.0.1
